fix: delete app cache when cancel installation
