+++ pkg/appinstaller/helm.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"flag"
 	"fmt"
+	"net/http"
 	"net/http/httputil"
 	"os"
 	"path/filepath"
@@ -46,6 +47,7 @@ import (
 
 var (
 	systemServerHost = ""
+	middlewareTypes  = []string{tapr.TypePostgreSQL.String(), tapr.TypeMongoDB.String(), tapr.TypeRedis.String(), tapr.TypeNats.String()}
 )
 
 func init() {
@@ -649,6 +651,9 @@ func (h *HelmOps) Uninstall() error {
 			}
 		}
 	}
+
+	appCacheDirs, _ := utils.TryToGetAppdataDirFromDeployment(context.TODO(), h.app.Namespace, h.app.AppName, h.app.OwnerName)
+
 	err = helm.UninstallCharts(h.actionConfig, h.app.AppName)
 	if err != nil {
 		return err
@@ -656,6 +661,46 @@ func (h *HelmOps) Uninstall() error {
 	err = h.unregisterAppPerm()
 	if err != nil {
 		klog.Errorf("Failed to unregister app err=%v", err)
+	}
+
+	// delete middleware requests crd
+	namespace := fmt.Sprintf("%s-%s", "user-system", h.app.OwnerName)
+	for _, mt := range middlewareTypes {
+		name := fmt.Sprintf("%s-%s", h.app.AppName, mt)
+		err = tapr.DeleteMiddlewareRequest(context.TODO(), h.kubeConfig, namespace, name)
+		if err != nil && !apierrors.IsNotFound(err) {
+			klog.Errorf("Failed to delete middleware request namespace=%s name=%s err=%v", namespace, name, err)
+		}
+	}
+
+	if len(appCacheDirs) > 0 {
+		terminusNonce, e := utils.GenTerminusNonce()
+		if e != nil {
+			klog.Errorf("Failed to generate terminus nonce err=%v", e)
+		} else {
+			c := resty.New().SetTimeout(2*time.Second).
+				SetHeader(constants.AuthorizationTokenKey, h.token).
+				SetHeader("Terminus-Nonce", terminusNonce)
+			nodes, e := client.CoreV1().Nodes().List(context.TODO(), metav1.ListOptions{})
+			if e == nil {
+				for _, dir := range appCacheDirs {
+					for _, n := range nodes.Items {
+						URL := fmt.Sprintf(constants.AppDataDirURL, h.app.OwnerName, dir)
+						c.SetHeader("X-Terminus-Node", n.Name)
+						c.SetHeader("x-bfl-user", h.app.OwnerName)
+						res, e := c.R().Delete(URL)
+						if e != nil {
+							klog.Errorf("Failed to delete dir err=%v", e)
+						}
+						if res.StatusCode() != http.StatusOK {
+							klog.Infof("delete app cache failed with: %v", res.String())
+						}
+					}
+				}
+			} else {
+				klog.Error("Failed to get nodes err=%v", e)
+			}
+		}
 	}
 
 	if !utils.IsProtectedNamespace(h.app.Namespace) {
