+++ pkg/utils/app.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"net"
 	"net/netip"
+	"path/filepath"
 	"strings"
 	"time"
 
@@ -18,6 +19,7 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/util/retry"
 	"k8s.io/klog/v2"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -501,4 +503,75 @@ func parseDestination(dest string) (string, string, error) {
 	}
 
 	return alias, tokens[len(tokens)-1], nil
+}
+
+func TryToGetAppdataDirFromDeployment(ctx context.Context, namespace, name, owner string) (dirs []string, err error) {
+	userspaceNs := UserspaceName(owner)
+	config, err := ctrl.GetConfig()
+	if err != nil {
+		return dirs, err
+	}
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return dirs, err
+	}
+	sts, err := clientset.AppsV1().StatefulSets(userspaceNs).Get(ctx, "bfl", metav1.GetOptions{})
+	if err != nil {
+		return dirs, err
+	}
+	appName := fmt.Sprintf("%s-%s", namespace, name)
+	appCachePath := sts.GetAnnotations()["appcache_hostpath"]
+	if len(appCachePath) == 0 {
+		return dirs, errors.New("empty appcache_hostpath")
+	}
+	if !strings.HasSuffix(appCachePath, "/") {
+		appCachePath += "/"
+	}
+	dClient, err := versioned.NewForConfig(config)
+	if err != nil {
+		return dirs, err
+	}
+	appCRD, err := dClient.AppV1alpha1().Applications().Get(ctx, appName, metav1.GetOptions{})
+	if err != nil {
+		return dirs, err
+	}
+	deploymentName := appCRD.Spec.DeploymentName
+	deployment, err := clientset.AppsV1().Deployments(namespace).
+		Get(context.Background(), deploymentName, metav1.GetOptions{})
+	if err != nil {
+		if apierrors.IsNotFound(err) {
+			return tryToGetAppdataDirFromSts(ctx, namespace, deploymentName, appCachePath)
+		}
+		return dirs, err
+	}
+
+	for _, v := range deployment.Spec.Template.Spec.Volumes {
+		if v.HostPath != nil && strings.HasPrefix(v.HostPath.Path, appCachePath) && len(v.HostPath.Path) > len(appCachePath) {
+			dirs = append(dirs, filepath.Base(v.HostPath.Path))
+		}
+	}
+	return dirs, nil
+}
+
+func tryToGetAppdataDirFromSts(ctx context.Context, namespace, stsName, baseDir string) (dirs []string, err error) {
+	config, err := ctrl.GetConfig()
+	if err != nil {
+		return dirs, err
+	}
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return dirs, err
+	}
+
+	sts, err := clientset.AppsV1().StatefulSets(namespace).
+		Get(ctx, stsName, metav1.GetOptions{})
+	if err != nil {
+		return dirs, err
+	}
+	for _, v := range sts.Spec.Template.Spec.Volumes {
+		if v.HostPath != nil && strings.HasPrefix(v.HostPath.Path, baseDir) && len(v.HostPath.Path) > len(baseDir) {
+			dirs = append(dirs, filepath.Base(v.HostPath.Path))
+		}
+	}
+	return dirs, nil
 }
